mode activity fixtures to conftest for reuse
