+++ tests/integration/test_activity_repository.py
@@ -1,10 +1,4 @@
 import pytest
-from src.models.activity import Activity, ActivityLaps, ActivityDogs, ActivityCreate, ActivityDogsCreate
-from src.models.dog import Dog
-from src.models.runner import Runner
-from src.models.sport import Sport, SportType
-from src.models.kennel import Kennel
-from src.models.weather import Weather
 from src.repositories.activity_repository import activity_repository
 from datetime import date, timezone, datetime, timedelta
 from psycopg2.extras import RealDictCursor
@@ -13,99 +7,6 @@ from psycopg2.extras import RealDictCursor
 def activity_repo(test_db_conn):
     print(test_db_conn)
     return activity_repository(test_db_conn)
-
-@pytest.fixture()
-def test_kennel():
-    return Kennel(name = "Test Kennel")
-
-@pytest.fixture()
-def test_activity():
-    test_activity = Activity(
-        id=None,
-        timestamp=datetime(2025, 4, 1, 9, 30, tzinfo=timezone.utc),
-        runner=Runner(id=2, name="Obelix", kennel=Kennel(id=1, name="Les Gaulois")),
-        sport=Sport(id=1, name="Canicross", type = SportType.DRYLAND),
-        location="Forest Loop",
-        distance=8.0,
-        workout=True,
-        notes="Morning speed workout",
-        speed=20.3,
-        dogs=[
-            ActivityDogs(
-                    id=None,
-                    dog=Dog(
-                        id=1,
-                        name="Milou",
-                        breed="Terrier",
-                        date_of_birth=date(2023, 1, 1),
-                        kennel=Kennel(id=1, name="Les Gaulois")
-                    ),
-                    rating=9
-                ),
-                ActivityDogs(
-                    id=None,
-                    dog=Dog(
-                        id=2,
-                        name="Fido",
-                        breed="Golden Retriever",
-                        date_of_birth=date(2020, 7, 1),
-                        kennel=Kennel(id=1, name="Les Gaulois")
-                    ),
-                    rating=8
-                )
-        ],
-        laps=[
-            ActivityLaps(lap_number=1, speed=21.0, pace="02:51", lap_distance = 1, lap_time_delta=timedelta(minutes = 2, seconds = 51) ),
-            ActivityLaps(lap_number=2, speed=20.1, pace="02:59", lap_distance = 1, lap_time_delta=timedelta(minutes = 2, seconds = 59)),
-            ActivityLaps(lap_number=3, speed=19.8, pace="03:01", lap_distance = 1, lap_time_delta=timedelta(minutes = 3, seconds = 1))
-        ],
-        weather=Weather(
-            temperature=9.5,
-            humidity=0.85,
-            condition = "rainy"
-        )
-
-    )
-    return test_activity
-
-@pytest.fixture()
-def test_activity_create():
-    test_activity = ActivityCreate(
-        id=None,
-        timestamp=datetime(2025, 4, 1, 9, 30, tzinfo=timezone.utc),
-        runner_id=2,
-        sport_id=1,
-        location="Forest Loop",
-        distance=8.0,
-        workout=True,
-        speed=20.3,
-        dogs=[
-            ActivityDogsCreate(
-            id=None,
-            dog_id=1,
-            rating=9
-            ),
-            ActivityDogsCreate(
-                id=None,
-                dog_id = 2,
-                rating=8
-            )
-        ],
-        laps=[
-            ActivityLaps(lap_number=1, speed=21.0, pace="02:51", lap_distance = 1, lap_time_delta=timedelta(minutes = 2, seconds = 51) ),
-            ActivityLaps(lap_number=2, speed=20.1, pace="02:59", lap_distance = 1, lap_time_delta=timedelta(minutes = 2, seconds = 59)),
-            ActivityLaps(lap_number=3, speed=19.8, pace="03:01", lap_distance = 1, lap_time_delta=timedelta(minutes = 3, seconds = 1))
-        ],
-        weather=Weather(
-            temperature=9.5,
-            humidity=0.85,
-            condition = "rainy"
-        )
-
-    )
-    return test_activity
-
-
 
 def test_get_by_id(activity_repo):
     activity = activity_repo.get_by_id(1)
